Make linear indexing full range

+++ src/uTensor/core/types.hpp
@@ -22,7 +22,7 @@ class TensorShape {
   bool operator == (const TensorShape& other);
   bool operator != (const TensorShape& other);
   void update_dims();
-  uint16_t get_linear_size() const;
+  uint32_t get_linear_size() const;
   uint8_t num_dims() const;
   uint32_t linear_index(uint16_t i, uint16_t j, uint16_t k, uint16_t l) const;
   uint32_t num_elems() const;
